requirements.txt
joblib~=1.4.2
scikit_learn==1.6.1
pandas~=2.2.3
numpy~=2.2.4
scikit-learn~=1.6.1
requests~=2.32.3
PyYAML~=6.0.2
mlflow-skinny~=2.21.2